refactor: service 관련 상수 관련 파일 util 로 이동
